Fix uuid generator for windows

+++ ccoip/src/uuid_utils.cpp
@@ -5,12 +5,12 @@
 
 std::random_device rd;
 std::mt19937_64 generator(rd());
-std::uniform_int_distribution<uint8_t> dist(0, 255);
+std::uniform_int_distribution<uint32_t> dist(0, 255);
 
 void ccoip::uuid_utils::generate_uuid(ccoip_uuid &uuid_out) {
     // Generate 16 random bytes
     for (int i = 0; i < 16; ++i) {
-        uuid_out[i] = dist(generator);
+        uuid_out[i] = static_cast<uint8_t>(dist(generator));
     }
 
     // Set version to 4 (UUIDv4)
